src/requirements.txt
Django==3.0
djangorestframework==3.10.3
psycopg2==2.8.4
requests==2.22.0
